requirements.txt
rdkit-pypi
torch
numpy
pandas
tqdm
